fix: decouple reporter from app's runner graph and fix reporter types
Type-checking @packages/reporter pulled 20 errors out of @packages/app's
source. Reporter's program reached app's entire runner subtree and then
checked it under reporter's config (ES2018, no @cy/* paths, no vue
shims) instead of app's, so app code that is correct under its own
vue-tsc check-ts reported as broken.

The reach came from a single type import: mobx-runner-store pulled
AutomationStatus from the store barrel, which re-exports aut-store,
runner-ui-store and the runner subtree behind them. The automation
constant and its keyof type need none of that, so they move to their own
module. The store barrel re-exports them, keeping the existing surface
for event-manager and anything else importing from '../store'.

Reporter's two e2e specs that imported MobxRunnerStore from the barrel
now import the module directly. Together this takes app out of
reporter's program entirely, and @packages/app check-ts still passes.

Also fixes the type errors this exposed in reporter itself:

- Log widened aliasType to string, discarding the union InstrumentProps
  declares. Session did the same to tagType, discarding what
  determineTagType returns. Both now keep the precise type, and tag owns
  the TagType vocabulary its consumers conform to, so a missing member
  is a compile error rather than a silent widening.
- TestError already returned null for a missing err at runtime, but
  typed the prop as required.
- IconCypressStudio takes fill colors, not stroke. React dropped the
  unknown strokeColor attribute, so removing it keeps rendering as-is.
- Reporter now references the driver's internal-types-lite, as app does,
  for the internal Cypress.state and backendRequestHandler APIs.
