+++ packages/app/src/store/automation-status.ts
@@ -1,0 +1,8 @@
+export const automation = {
+  CONNECTING: 'CONNECTING',
+  MISSING: 'MISSING',
+  CONNECTED: 'CONNECTED',
+  DISCONNECTED: 'DISCONNECTED',
+} as const
+
+export type AutomationStatus = keyof typeof automation
