+++ packages/app/src/store/runner-ui-store.ts
@@ -1,14 +1,7 @@
 import { defineStore } from 'pinia'
 import { runnerConstants } from '../runner/runner-constants'
-
-export const automation = {
-  CONNECTING: 'CONNECTING',
-  MISSING: 'MISSING',
-  CONNECTED: 'CONNECTED',
-  DISCONNECTED: 'DISCONNECTED',
-} as const
-
-export type AutomationStatus = keyof typeof automation
+import { automation } from './automation-status'
+import type { AutomationStatus } from './automation-status'
 
 /**
  * Store for reactive properties used in the runner UI.
