+++ packages/reporter/cypress/e2e/runnables.cy.ts
@@ -4,7 +4,7 @@ import { itHandlesFileOpening } from '../support/utils'
 import type { BaseReporterProps } from '../../src/main'
 import type { RunnablesErrorModel } from '../../src/runnables/runnable-error'
 import appState from '../../src/lib/app-state'
-import { MobxRunnerStore } from '@packages/app/src/store'
+import { MobxRunnerStore } from '@packages/app/src/store/mobx-runner-store'
 import scroller from '../../src/lib/scroller'
 
 const runnerStore = new MobxRunnerStore('e2e')
