+++ packages/reporter/index.d.ts
@@ -1,3 +1,5 @@
+/// <reference path="../driver/types/internal-types-lite.d.ts" />
+
 /// <reference path="../../cli/types/cy-blob-util.d.ts" />
 /// <reference path="../../cli/types/cy-bluebird.d.ts" />
 /// <reference path="../../cli/types/cy-minimatch.d.ts" />
@@ -17,7 +19,7 @@ declare namespace Cypress {
   }
 }
 
-declare module "*.svg" {
-  const content: any;
-  export default content;
+declare module '*.svg' {
+  const content: any
+  export default content
 }
