+++ packages/reporter/src/lib/tag.tsx
@@ -2,8 +2,10 @@ import cs from 'classnames'
 import React from 'react'
 import Tooltip from './tooltip'
 
+export type TagType = 'agent' | 'count' | 'dom' | 'failed-status' | 'primitive' | 'route' | 'successful-status' | 'warned-status'
+
 interface TagProps {
-  type?: 'agent' | 'count' | 'dom' | 'failed-status' | 'primitive' | 'route' | 'successful-status' | 'warned-status'
+  type?: TagType
   content: React.ReactNode | string
   count?: number
   tooltipMessage?: React.ReactNode | string
